feat: add opt-in tradingview native market chart
