+++ packages/kit/src/components/TradingView/TradingViewNative/types.ts
@@ -1,4 +1,9 @@
-import type { ITradingViewV2Props } from '../TradingViewV2';
-import type { WebViewProps } from 'react-native-webview';
-
-export type ITradingViewNativeProps = ITradingViewV2Props & WebViewProps;
+export interface ITradingViewNativeProps {
+  testID?: string;
+  networkId?: string;
+  tokenAddress?: string;
+  symbol?: string;
+  decimal?: number;
+  nativeControlsLayoutMode?: 'mobile' | 'desktop';
+  onNativeSubIndicatorCountChange?: (count: number | null) => void;
+}
