+++ packages/kit/src/views/Market/MarketDetailV2/hooks/index.ts
@@ -4,3 +4,4 @@ export * from './useMarketHolders';
 export * from './useAutoRefreshTokenDetail';
 export * from './useBtcMetadata';
 export * from './BtcMetadataContext';
+export * from './useTradingViewNativeInMarketDetail';
