+++ packages/kit/src/views/Market/MarketDetailV2/hooks/useTradingViewNativeInMarketDetail.ts
@@ -1,0 +1,10 @@
+import { useDevSettingsPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms/devSettings';
+
+export function useTradingViewNativeInMarketDetail() {
+  const [devSettings] = useDevSettingsPersistAtom();
+
+  return Boolean(
+    devSettings.enabled &&
+    devSettings.settings?.useTradingViewNativeInMarketDetail,
+  );
+}
